Turn query type into a context
