+++ decidim-api/spec/types/translated_field_type_spec.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 require "spec_helper"
-require "decidim/api/test/type_helpers"
+require "decidim/api/test/type_context"
 
 module Decidim
   module Api
     describe TranslatedFieldType do
-      include Test::TypeHelpers
+      include_context "graphql type"
 
       let(:model) do
         {
