+++ decidim-core/spec/types/process_type_spec.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 require "spec_helper"
-require "decidim/api/test/type_helpers"
+require "decidim/api/test/type_context"
 
 module Decidim
-  describe ProcessType do
-    include Decidim::Api::Test::TypeHelpers
+  describe ProcessType, type: :graphql do
+    include_context "graphql type"
 
     let(:model) { create(:participatory_process) }
